+++ R/subvariables.R
@@ -65,7 +65,7 @@ setMethod("subvariables<-", c("CategoricalArrayVariable", "Subvariables"),
             body <- list(subvariables=new)
             payload <- toJSON(body)
             crPATCH(self(x), body=payload)
-            dropCache(absoluteURL("./cube/", datasetReference(x)))
+            dropCache(cubeURL(x))
             tuple(x)$subvariables <- new
         }
         return(x)