+++ R/variable-summary.R
@@ -3,11 +3,7 @@ getSummary <- function (x) {
     if (is.null(url)) {
         halt("No summary available")
     }
-    out <- crGET(url)
-    ## Summaries don't return as shoji entities
-    # if (!is.shoji(url)) {
-    #     halt("Error in retrieving summary")
-    # }
+    out <- crGET(url, query=list(filter=toJSON(zcl(activeFilter(x)))))
     if (is.Datetime(x)) {
         toR <- columnParser("datetime")
         for (i in c("min", "max")) out[[i]] <- toR(out[[i]])
@@ -46,9 +42,8 @@ table <- function (..., exclude, useNA=c("no", "ifany", "always"), dnn, deparse.
             query=toJSON(query),
             filter=filters[1]
         )
-        cube_url <- absoluteURL("./cube/", datasetReference(dots[[1]]))
-        cube <- CrunchCube(crGET(cube_url, query=query),
-            useNA=match.arg(useNA))
+        cube_url <- cubeURL(dots[[1]])
+        cube <- CrunchCube(crGET(cube_url, query=query), useNA=match.arg(useNA))
         return(as.table(as.array(cube)))
     } else if (any(are.vars)) {
         halt("Cannot currently tabulate Crunch variables with ",