+++ tests/testthat/test-crunchbox.R
@@ -12,18 +12,20 @@ with_mock_HTTP({
     ds3 <- loadDataset("ECON.sav")
 
     test_that("Basic box", {
-        expect_POST(crunchBox(ds),
+        expect_POST(crunchBox(ds, filters=NULL),
             '/api/datasets/dataset1/boxdata/',
             '{"element":"shoji:entity","body":{"filters":[]}}')
     })
-    test_that("Basic box with metadata", {
+    test_that("Basic box with metadata and filters", {
         expect_POST(crunchBox(ds, title="Test box"),
             '/api/datasets/dataset1/boxdata/',
-            '{"element":"shoji:entity","body":{"filters":[],',
+            '{"element":"shoji:entity","body":{"filters":[',
+            '{"filter":"/api/datasets/dataset1/filters/filter1/"},',
+            '{"filter":"/api/datasets/dataset1/filters/filter2/"}],',
             '"title":"Test box"}}')
     })
     test_that("Select variables in box", {
-        expect_POST(crunchBox(ds[2:4]),
+        expect_POST(crunchBox(ds[2:4], filters=NULL),
             '/api/datasets/dataset1/boxdata/',
             '{"element":"shoji:entity","body":{"filters":[],',
             '"where":{"function":"select","args":[{"map":{',
@@ -45,7 +47,16 @@ with_mock_HTTP({
             '{"variable":"/api/datasets/dataset3/variables/starttime/"}}}]}}}')
     })
     test_that("Select filters in box", {
-
+        expect_POST(crunchBox(ds, filters=filters(ds)[c(2, 1)], title="Test box"),
+            '/api/datasets/dataset1/boxdata/',
+            '{"element":"shoji:entity","body":{"filters":[',
+            '{"filter":"/api/datasets/dataset1/filters/filter2/"},',
+            '{"filter":"/api/datasets/dataset1/filters/filter1/"}],',
+            '"title":"Test box"}}')
+        expect_POST(crunchBox(ds, filters=filters(ds)[2]),
+            '/api/datasets/dataset1/boxdata/',
+            '{"element":"shoji:entity","body":{"filters":[',
+            '{"filter":"/api/datasets/dataset1/filters/filter2/"}]}}')
     })
     test_that("Input validation", {
         expect_error(crunchBox(), "'dataset' must be a CrunchDataset")
@@ -56,10 +67,10 @@ with_mock_HTTP({
     })
     test_that("Box too big message", {
         with_mock(`crunch:::.boxlimit`=function () -1, {
-            expect_error(crunchBox(ds[2:5]),
-                "4 variables and 0 filters results in too many cubes")
+            expect_error(crunchBox(ds[2:5], filters=filters(ds)[1]),
+                "4 variables and 1 filter results in too many cubes")
             expect_error(crunchBox(ds[2]),
-                "1 variable and 0 filters results in too many cubes")
+                "1 variable and 2 filters results in too many cubes")
         })
     })
 })