+++ tests/testthat/test-misc.R
@@ -9,18 +9,6 @@ test_that("is.error", {
     expect_error("not an error", NA)
 })
 
-test_that("selectFrom selects what it should", {
-    l1 <- list(list(a=1, b=2), list(c=3, b=4))
-    expect_identical(selectFrom("b", l1), c(2, 4))
-    expect_identical(selectFrom("a", l1), c(1, NA))
-    expect_identical(selectFrom("a", l1, ifnot=4), c(1, 4))
-    expect_identical(selectFrom("d", l1), c(NA, NA))
-    l2 <- l1
-    l2[[2]] <- 4
-    expect_identical(selectFrom("b", l2), c(2, NA))
-    expect_error(selectFrom("b", 5), "xlist must be a list object")
-})
-
 test_that("rethrow a caught error", {
     e <- try(halt("error in a box"), silent=TRUE)
     expect_true(is.error(e))