+++ tests/testthat/test-get-datasets.R
@@ -10,10 +10,10 @@ with_mock_HTTP({
 
     test_that("loadDataset loads", {
         ## NOTE: implementing the active dataset loading only first
-        ds <- try(loadDataset("test ds"))
+        ds <- loadDataset("test ds")
         expect_true(is.dataset(ds))
         expect_identical(name(ds), "test ds")
-        ds <- try(loadDataset(2))
+        ds <- loadDataset(2)
         expect_true(is.dataset(ds))
         expect_identical(name(ds), "test ds")
         expect_error(loadDataset(666), "subscript out of bounds")
@@ -36,75 +36,73 @@ with_mock_HTTP({
     })
 })
 
-if (run.integration.tests) {
-    with_test_authentication({
-        test_that("datasetCatalog gets what we expect", {
-            col0 <- datasets()
-            expect_is(col0, "DatasetCatalog")
-            with(test.dataset(), {
-                col1 <- datasets()
-                expect_is(col1, "DatasetCatalog")
-                expect_equal(length(col1), length(col0) + 1)
-            })
-        })
+with_test_authentication({
+    test_that("datasets() gets the dataset catalog", {
+        col0 <- datasets()
+        expect_is(col0, "DatasetCatalog")
         with(test.dataset(), {
-            dsname <- name(ds)
-            test_that("Dataset list can be retrieved if authenticated", {
-                expect_true(is.character(listDatasets()))
-                expect_true(length(listDatasets())>0)
-                expect_true(is.character(dsname))
-                expect_true(nchar(dsname) > 0)
-                expect_true(dsname %in% listDatasets())
-            })
-
-            test_that("A dataset object can be retrieved, if it exists", {
-                expect_true(is.dataset(loadDataset(dsname)))
-                dsnum <- which(listDatasets() %in% dsname)
-                expect_true(is.numeric(dsnum))
-                expect_true(is.dataset(loadDataset(dsnum)))
-            })
-
-            test_that("deleteDataset by name", {
-                out <- try(deleteDataset(dsname))
-                expect_false(dsname %in% listDatasets())
-            })
-        })
-
-        with(test.dataset(), {
-            dsname <- name(ds)
-            dsnum <- which(listDatasets() %in% dsname)
-            test_that("deleteDataset by index", {
-                expect_true(dsname %in% listDatasets())
-                out <- deleteDataset(dsnum)
-                expect_false(dsname %in% listDatasets())
-            })
-        })
-
-        with(test.dataset(), {
-            dsname <- name(ds)
-            test_that("deleteDataset on Dataset object", {
-                expect_true(dsname %in% listDatasets())
-                out <- deleteDataset(ds)
-                expect_false(dsname %in% listDatasets())
-            })
-        })
-
-        with(test.dataset(), {
-            dsname <- name(ds)
-            newname <- paste0("New name ", now())
-
-            test_that("renaming a dataset refreshes the dataset list", {
-                expect_true(dsname %in% listDatasets())
-                name(ds) <- newname
-                expect_false(dsname %in% listDatasets())
-                expect_true(newname %in% listDatasets())
-            })
-
-            test_that("deleting a dataset refreshes the dataset list", {
-                delete(ds)
-                expect_false(dsname %in% listDatasets())
-                expect_false(newname %in% listDatasets())
-            })
+            col1 <- datasets()
+            expect_is(col1, "DatasetCatalog")
+            expect_equal(length(col1), length(col0) + 1)
         })
     })
-}
+    with(test.dataset(), {
+        dsname <- name(ds)
+        test_that("Dataset list can be retrieved if authenticated", {
+            expect_true(is.character(listDatasets()))
+            expect_true(length(listDatasets())>0)
+            expect_true(is.character(dsname))
+            expect_true(nchar(dsname) > 0)
+            expect_true(dsname %in% listDatasets())
+        })
+
+        test_that("A dataset object can be retrieved, if it exists", {
+            expect_true(is.dataset(loadDataset(dsname)))
+            dsnum <- which(listDatasets() %in% dsname)
+            expect_true(is.numeric(dsnum))
+            expect_true(is.dataset(loadDataset(dsnum)))
+        })
+
+        test_that("deleteDataset by name", {
+            out <- try(deleteDataset(dsname))
+            expect_false(dsname %in% listDatasets())
+        })
+    })
+
+    with(test.dataset(), {
+        dsname <- name(ds)
+        dsnum <- which(listDatasets() %in% dsname)
+        test_that("deleteDataset by index", {
+            expect_true(dsname %in% listDatasets())
+            out <- deleteDataset(dsnum)
+            expect_false(dsname %in% listDatasets())
+        })
+    })
+
+    with(test.dataset(), {
+        dsname <- name(ds)
+        test_that("deleteDataset on Dataset object", {
+            expect_true(dsname %in% listDatasets())
+            out <- deleteDataset(ds)
+            expect_false(dsname %in% listDatasets())
+        })
+    })
+
+    with(test.dataset(), {
+        dsname <- name(ds)
+        newname <- paste0("New name ", now())
+
+        test_that("renaming a dataset refreshes the dataset list", {
+            expect_true(dsname %in% listDatasets())
+            name(ds) <- newname
+            expect_false(dsname %in% listDatasets())
+            expect_true(newname %in% listDatasets())
+        })
+
+        test_that("deleting a dataset refreshes the dataset list", {
+            delete(ds)
+            expect_false(dsname %in% listDatasets())
+            expect_false(newname %in% listDatasets())
+        })
+    })
+})