+++ NEWS.md
@@ -1,6 +1,7 @@
 ### crunch 1.13.7 (under development)
-* `flattenOrder` to strip out nested groups from an order
+* `crunchBox` to make a public, embeddable analysis widget
 * `settings` and `settings<-` to view and modify dataset-level controls, such as default "weight" and viewer permissions ("viewers_can_change_weight", "viewers_can_export")
+* `flattenOrder` to strip out nested groups from an order
 * Univariate statistics on variables, such as `mean`, `median`, and `sd`, now respect filter expressions, as does the `summary` method.
 * "median" can now be used in `crtabs`
 * Copying and deriving variables now bring in the "notes" attribute.